+++ bin/browser.js
@@ -37,11 +37,12 @@ const getOutput = async (page, request) => {
     return output.toString('base64');
 };
 
-const callChrome = async () => {
+const callChrome = async pup => {
     let browser;
     let page;
     let output;
     let remoteInstance;
+	const puppet = (pup || puppeteer);
 
     try {
         if (request.options.remoteInstanceUrl || request.options.browserWSEndpoint ) {
@@ -58,17 +59,22 @@ const callChrome = async () => {
             }
 
             try {
-                browser = await puppeteer.connect( options );
+                browser = await puppet.connect( options );
 
                 remoteInstance = true;
             } catch (exception) { /** does nothing. fallbacks to launching a chromium instance */}
         }
 
         if (!browser) {
-            browser = await puppeteer.launch({
+            browser = await puppet.launch({
                 ignoreHTTPSErrors: request.options.ignoreHttpsErrors,
                 executablePath: request.options.executablePath,
-                args: request.options.args || []
+                args: request.options.args || [],
+                pipe: request.options.pipe || false,
+                env: {
+                    ...(request.options.env || {}),
+                    ...process.env
+                },
             });
         }
 
@@ -128,13 +134,13 @@ const callChrome = async () => {
         }
 
         if (request.options && request.options.device) {
-            const devices = require('puppeteer/DeviceDescriptors');
+            const devices = puppeteer.devices;
             const device = devices[request.options.device];
             await page.emulate(device);
         }
 
         if (request.options && request.options.emulateMedia) {
-            await page.emulateMedia(request.options.emulateMedia);
+            await page.emulateMediaType(request.options.emulateMedia);
         }
 
         if (request.options && request.options.viewport) {
@@ -269,4 +275,8 @@ const callChrome = async () => {
     }
 };
 
-callChrome();
+if (require.main === module) {
+	callChrome();
+}
+
+exports.callChrome = callChrome;